file_system_docker/catalog_service/requirements.txt
grpcio
grpcio-tools
pymongo